Extract HAML helpers to separate module
