+++ lib/machinery.rb
@@ -103,6 +103,7 @@ require_relative "man_task"
 require_relative "comparison"
 require_relative "serve_html_task"
 require_relative "file_diff"
+require_relative "haml_helpers"
 require_relative "server"
 require_relative "html"
 require_relative "machinery_helper"
